requirements.txt
streamlit
pyyaml
langchain-ollama
langchain-core
langgraph
chromadb
python-telegram-bot
apscheduler
ddgs
playwright
beautifulsoup4
requests
python-docx
pypdf
openpyxl
python-pptx
ebooklib
Pillow
pytesseract
pdf2image
faster-whisper
pydantic
pydantic-settings
jinja2
streamlit
requests
python-dotenv
rich
ffmpeg-python